@@ -1,20 +1,30 @@
 {
-  "name": "p5js-react-template",
+  "name": "p5-react-template",
   "private": true,
   "version": "0.0.0",
+  "type": "module",
   "scripts": {
     "dev": "vite",
     "build": "vite build",
+    "lint": "eslint . --ext js,jsx --report-unused-disable-directives --max-warnings 0",
     "preview": "vite preview"
   },
   "dependencies": {
-    "p5": "^1.4.1",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2"
+    "leva": "^0.9.35",
+    "p5": "^1.8.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0"
   },
   "devDependencies": {
-    "@vitejs/plugin-react": "^1.0.7",
-    "sass": "^1.49.9",
-    "vite": "^2.8.0"
+    "@types/react": "^18.2.15",
+    "@types/react-dom": "^18.2.7",
+    "@vitejs/plugin-react-swc": "^3.3.2",
+    "eslint": "^8.45.0",
+    "eslint-plugin-react": "^7.32.2",
+    "eslint-plugin-react-hooks": "^4.6.0",
+    "eslint-plugin-react-refresh": "^0.4.3",
+    "standard": "^17.1.0",
+    "vite": "^4.4.5",
+    "vite-plugin-glsl": "^1.1.2"
   }
 }
