@@ -1,11 +1,9 @@
-import React from "react"
-import ReactDOM from "react-dom"
-import "./assets/styles/index.scss"
-import App from "./App"
+import React from 'react'
+import ReactDOM from 'react-dom/client'
+import App from './App.jsx'
 
-ReactDOM.render(
+ReactDOM.createRoot(document.getElementById('root')).render(
   <React.StrictMode>
     <App />
-  </React.StrictMode>,
-  document.getElementById("root")
+  </React.StrictMode>
 )
