@@ -1,30 +1,13 @@
-import { useState, useEffect } from 'react'
-import Canvas, {moveX} from './sketch/canvas/Canvas'
-import moveShape from './sketch/functions/MoveShape'
-import Panel from './components/Panel'
-import Button from './components/Button'
+import { useLevaControls } from './settings'
+import Canvas from './sketch/canvas/Canvas'
 
-function App() {
+function App () {
+  const { data, frameRateRef } = useLevaControls()
 
   return (
-    <div className="App">
-      <Panel>
-        <Button
-          text="Left"
-          onClick={() => moveX(-10)}
-        />
-        <Button
-          text="Right"
-          onClick={() => moveX(10)}
-        />
-      </Panel>
-      <Canvas
-        sizeX={200}
-        sizeY={200}
-        fullpage={true}
-        centerMode={true}
-      />
-    </div>
+    <>
+      <Canvas settings={{ ...data, frameRateRef }} />
+    </>
   )
 }
 
